requirements.txt
google-genai
requests
winsdk
pycaw
comtypes
psutil
cryptography
google-auth 
google-auth-oauthlib 
google-api-python-client
icalendar
Pillow
numpy
opencv-python
mss